Neatened the user page, added routes and removed the reset button
