+++ client/src/page/teacherAbsences.css
@@ -1,0 +1,23 @@
+
+.teacherAbsences-week {
+    display: flex;
+    flex-direction: column;
+
+    align-items: center;
+    
+}
+
+.teacherAbsenceForm {
+    display: flex;
+    flex-direction: column;
+
+    align-items: center;
+    gap: 30px;
+
+}
+
+.teacherAbsences-button {
+
+    display: flex;
+    gap: 40px
+}
